build error fixes

+++ src/models/coordinates.cairo
@@ -1,5 +1,5 @@
 #[derive(Drop, Copy, Serde, Introspect)]
 pub struct Coordinates {
-    x: u8,
-    y: u8
+    x: u32,
+    y: u32
 }
